pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Mirdalan <mirdalan.p@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
flask = "^2.2.2"
gunicorn = "^20.1.0"
flake8 = "^6.0.0"
python-dotenv = "^0.21.1"
pytest = "^7.2.1"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
